+++ test/support/postgres/migrations/05_migrate_lotus_v6.exs
@@ -1,0 +1,11 @@
+defmodule Lotus.Test.Repo.Migrations.MigrateLotusV6 do
+  use Ecto.Migration
+
+  def up do
+    Lotus.Migrations.up()
+  end
+
+  def down do
+    Lotus.Migrations.down()
+  end
+end